requirements.txt
fastapi
uvicorn
pydantic
openai-whisper
python-multipart
python-dotenv
